Chnaged the implementations
