+++ connectid/urls.py
@@ -19,7 +19,6 @@ from django.urls import include, path
 urlpatterns = [
     path('users/', include('users.urls')),
     path('messaging/', include('messaging.urls')),
-    path('connectmessaging/', include('connectmessaging.urls')),
     path('admin/', admin.site.urls),
     path('o/', include('oauth2_provider.urls', namespace='oauth2_provider')),
 ]
